e2e: split send mention helper out so not stuck in individual test file
